server/pyproject.toml
[tool.poetry]
name = "baza_store"
version = "0.1.0"
description = ""
authors = ["iizhukov <iizhukov@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"
requests = "^2.31.0"
dadata = "^21.10.1"
text-unidecode = "^1.3"
python-slugify = "^8.0.4"
pillow = "^10.2.0"
drf-spectacular = "^0.27.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
runserver = "scripts.runserver:main"
migrate = "scripts.migrate:main"
